examples/requirements.txt
langchain
langchain-openai
python-dotenv
browser-use
playwright
lxml
lxml-html-clean
